review notes applied
