+++ test/org/mockitousage/junitrule/RuleTestWithFactoryMethodTest.java
@@ -1,9 +1,11 @@
-package org.mockito.rules;
+package org.mockitousage.junitrule;
 
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.MethodRule;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.rules.MockitoJUnit;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -11,7 +13,7 @@ import static org.junit.Assert.assertNotNull;
 public class RuleTestWithFactoryMethodTest {
 
     @Rule
-    public MockitoJUnitRule mockitoJUnitRule = MockitoJUnitRule.rule();
+    public MethodRule mockitoJUnitRule = MockitoJUnit.rule();
 
     @Mock
     private Injected injected;
